requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchainhub
langchain-groq
huggingface-hub
sentence-transformers
chromadb
PyMuPDF